requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.1.0
fastapi==0.115.6
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.1
python-dotenv==1.0.0
PyYAML==6.0.3
six==1.16.0
SQLAlchemy==2.0.45
starlette==0.41.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.34.0
watchfiles==1.1.1
websockets==15.0.1
